+++ CHANGES.md
@@ -5,6 +5,7 @@ Changes in 8.2
 --------------
 - Fix replace regex: remove slash & allow space in FileUpload.fnc.php
 - Always Use Last Year's Picture if Missing in PrintClassLists.php
+- Fix #329 SQL error division by zero in t_update_mp_stats(): set min Credits to 1 in Courses.fnc.php
 
 Changes in 8.1.1
 ----------------